requirements.txt
certifi==2018.4.16
dj-database-url==0.5.0
Django==2.0.7
gunicorn==19.9.0
pipenv==2018.7.1
psycopg2==2.7.5
pytz==2018.5
reverse==0.1.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
psycopg2-binary==2.7.5
whitenoise==3.2
model-mommy==1.6.0
pagseguro==0.3.2
django-paypal==0.5.0
Pillow==5.2.0
easy-thumbnails==2.5
django-watson==1.5.2